+++ enterprise_data/api/v1/urls.py
@@ -10,6 +10,10 @@ from django.urls import re_path
 from enterprise_data.api.v1.views import enterprise_admin as enterprise_admin_views
 from enterprise_data.api.v1.views import enterprise_learner as enterprise_learner_views
 from enterprise_data.api.v1.views import enterprise_offers as enterprise_offers_views
+from enterprise_data.api.v1.views.analytics_enrollments import (
+    AdvanceAnalyticsEnrollmentStatsView,
+    AdvanceAnalyticsIndividualEnrollmentsView,
+)
 from enterprise_data.constants import UUID4_REGEX
 
 app_name = 'enterprise_data_api_v1'
@@ -35,6 +39,11 @@ router.register(
     enterprise_learner_views.EnterpriseLearnerCompletedCoursesViewSet,
     'enterprise-learner-completed-courses',
 )
+router.register(
+    r'enterprise/(?P<enterprise_id>.+)/module-performance',
+    enterprise_admin_views.EnterpriseExecEdLCModulePerformanceViewSet,
+    'enterprise-admin-module-performance',
+)
 
 urlpatterns = [
     re_path(
@@ -46,6 +55,16 @@ urlpatterns = [
         fr'^admin/anlaytics/(?P<enterprise_id>{UUID4_REGEX})$',
         enterprise_admin_views.EnterpriseAdminAnalyticsAggregatesView.as_view(),
         name='enterprise-admin-analytics-aggregates'
+    ),
+    re_path(
+        fr'^admin/anlaytics/(?P<enterprise_uuid>{UUID4_REGEX})/enrollments/stats$',
+        AdvanceAnalyticsEnrollmentStatsView.as_view(),
+        name='enterprise-admin-analytics-enrollments-stats'
+    ),
+    re_path(
+        fr'^admin/anlaytics/(?P<enterprise_uuid>{UUID4_REGEX})/enrollments$',
+        AdvanceAnalyticsIndividualEnrollmentsView.as_view(),
+        name='enterprise-admin-analytics-enrollments'
     ),
     re_path(
         fr'^admin/anlaytics/(?P<enterprise_id>{UUID4_REGEX})/skills/stats',