+++ enterprise_data/api/v1/views/enterprise_admin.py
@@ -1,11 +1,11 @@
 """
 Views for enterprise admin api v1.
 """
-
 from datetime import datetime, timedelta
 
 from edx_rbac.decorators import permission_required
 from edx_rest_framework_extensions.auth.jwt.authentication import JwtAuthentication
+from rest_framework import filters, viewsets
 from rest_framework.response import Response
 from rest_framework.status import HTTP_200_OK, HTTP_404_NOT_FOUND
 from rest_framework.views import APIView
@@ -22,8 +22,14 @@ from enterprise_data.admin_analytics.utils import (
     get_top_skills_csv_data,
 )
 from enterprise_data.api.v1 import serializers
-from enterprise_data.models import EnterpriseAdminLearnerProgress, EnterpriseAdminSummarizeInsights
+from enterprise_data.models import (
+    EnterpriseAdminLearnerProgress,
+    EnterpriseAdminSummarizeInsights,
+    EnterpriseExecEdLCModulePerformance,
+)
 from enterprise_data.utils import date_filter
+
+from .base import EnterpriseViewSetMixin
 
 
 class EnterpriseAdminInsightsView(APIView):
@@ -109,16 +115,16 @@ class EnterpriseAdminAnalyticsAggregatesView(APIView):
         # 1. minimum enrollment date from the data as the start_date
         # 2. today's date as the end_date
         start_date = serializer.data.get(
-            "start_date", enrollment.enterprise_enrollment_date.min()
+            'start_date', enrollment.enterprise_enrollment_date.min()
         )
-        end_date = serializer.data.get("end_date", datetime.now())
+        end_date = serializer.data.get('end_date', datetime.now())
 
         # Date filtering.
         dff = date_filter(
             start=start_date,
             end=end_date,
             data_frame=enrollment.copy(),
-            date_column="enterprise_enrollment_date",
+            date_column='enterprise_enrollment_date',
         )
 
         enrolls = len(dff)
@@ -128,7 +134,7 @@ class EnterpriseAdminAnalyticsAggregatesView(APIView):
             start=start_date,
             end=end_date,
             data_frame=enrollment.copy(),
-            date_column="passed_date",
+            date_column='passed_date',
         )
 
         completions = dff.has_passed.sum()
@@ -138,7 +144,7 @@ class EnterpriseAdminAnalyticsAggregatesView(APIView):
             start=start_date,
             end=end_date,
             data_frame=engagement.copy(),
-            date_column="activity_date",
+            date_column='activity_date',
         )
 
         hours = round(dff.learning_time_seconds.sum() / 60 / 60, 1)
@@ -146,14 +152,14 @@ class EnterpriseAdminAnalyticsAggregatesView(APIView):
 
         return Response(
             data={
-                "enrolls": enrolls,
-                "courses": courses,
-                "completions": completions,
-                "hours": hours,
-                "sessions": sessions,
-                "last_updated_at": last_updated_at.date() if last_updated_at else None,
-                "min_enrollment_date": enrollment.enterprise_enrollment_date.min().date(),
-                "max_enrollment_date": enrollment.enterprise_enrollment_date.max().date(),
+                'enrolls': enrolls,
+                'courses': courses,
+                'completions': completions,
+                'hours': hours,
+                'sessions': sessions,
+                'last_updated_at': last_updated_at.date() if last_updated_at else None,
+                'min_enrollment_date': enrollment.enterprise_enrollment_date.min().date(),
+                'max_enrollment_date': enrollment.enterprise_enrollment_date.max().date(),
             },
             status=HTTP_200_OK,
         )
@@ -231,3 +237,25 @@ class EnterpriseAdminAnalyticsSkillsView(APIView):
         }
 
         return Response(data=response_data, status=HTTP_200_OK)
+
+
+class EnterpriseExecEdLCModulePerformanceViewSet(EnterpriseViewSetMixin, viewsets.ReadOnlyModelViewSet):
+    """
+    View to for getting enterprise exec ed learner module performance records.
+    """
+    serializer_class = serializers.EnterpriseExecEdLCModulePerformanceSerializer
+    filter_backends = (filters.OrderingFilter, filters.SearchFilter)
+    ordering_fields = '__all__'
+    ordering = ('last_access',)
+    search_fields = (
+        'username',
+        'course_name'
+    )
+
+    def get_queryset(self):
+        """
+        Return the queryset of EnterpriseExecEdLCModulePerformance objects.
+        """
+        return EnterpriseExecEdLCModulePerformance.objects.filter(
+            enterprise_customer_uuid=self.kwargs['enterprise_id'],
+        )